Made ClientSecret a secret and moved code into config
